--- a/pkg/storage/unified/sql/data/resource_version_batch_update.sql
+++ b/pkg/storage/unified/sql/data/resource_version_batch_update.sql
@@ -2,7 +2,9 @@ UPDATE {{ .Ident "resource_version" }}
 SET
     {{ .Ident "resource_version" }} = {{ .Arg .ResourceVersion }}
 WHERE 1 = 1
-    AND {{ .Ident "shard" }}    = {{ .Arg .Shard }}
     AND {{ .Ident "group" }}    = {{ .Arg .Group }}
     AND {{ .Ident "resource" }} = {{ .Arg .Resource }}
+    AND {{ .Ident "shard" }} IN (
+        {{ range $index, $element := .Shards }}{{ if $index }}, {{ end }}{{ .Arg $element }}{{ end }}
+    ) 
 ;