--- a/pkg/storage/unified/sql/data/resource_version_get.sql
+++ b/pkg/storage/unified/sql/data/resource_version_get.sql
@@ -1,5 +1,6 @@
 SELECT
-        {{ .Ident "resource_version" | .Into .ResourceVersion }}
+        {{ .Ident "resource_version" | .Into .Response.ResourceVersion }},
+        {{ .CurrentEpoch | .Into .Response.CurrentEpoch }}
     FROM {{ .Ident "resource_version" }}
     WHERE 1 = 1
         AND {{ .Ident "group" }}    = {{ .Arg .Group }}