--- a/pkg/storage/unified/sql/data/resource_version_inc.sql
+++ b/pkg/storage/unified/sql/data/resource_version_inc.sql
@@ -1,6 +1,6 @@
 UPDATE {{ .Ident "resource_version" }}
 SET
-    {{ .Ident "resource_version" }} = {{ .Arg .ResourceVersion}}
+    {{ .Ident "resource_version" }} = {{ .CurrentEpoch }}
 WHERE 1 = 1
     AND {{ .Ident "group" }}    = {{ .Arg .Group }}
     AND {{ .Ident "resource" }} = {{ .Arg .Resource }}