… pkg/storage/unified/sql/backend_test.go
… pkg/storage/unified/sql/data/resource_version_insert.sql
… pkg/storage/unified/sql/data/resource_version_lock.sql
… pkg/storage/unified/sql/data/resource_version_lock_shards.sql
… pkg/storage/unified/sql/data/resource_version_min.sql
… pkg/storage/unified/sql/data/resource_version_update.sql
… pkg/storage/unified/sql/db/migrations/resource_mig.go
… pkg/storage/unified/sql/queries.go
… pkg/storage/unified/sql/queries_test.go
… pkg/storage/unified/sql/sqltemplate/dialect.go
… pkg/storage/unified/sql/sqltemplate/dialect_mysql.go
… pkg/storage/unified/sql/sqltemplate/dialect_postgresql.go
… pkg/storage/unified/sql/sqltemplate/dialect_sqlite.go
… pkg/storage/unified/sql/sqltemplate/mocks/SQLTemplateIface.go
… pkg/storage/unified/sql/sqltemplate/mocks/WithResults.go
… pkg/storage/unified/sql/test/integration_test.go
… pkg/storage/unified/sql/testdata/mysql--resource_history_insert-insert into resource_history.sql
… pkg/storage/unified/sql/testdata/mysql--resource_read-without_resource_version.sql
… pkg/storage/unified/sql/testdata/mysql--resource_update-single path.sql
… pkg/storage/unified/sql/testdata/mysql--resource_version_get-single path.sql
… pkg/storage/unified/sql/testdata/mysql--resource_version_inc-increment resource version.sql
… pkg/storage/unified/sql/testdata/mysql--resource_version_insert-single path.sql
… pkg/storage/unified/sql/testdata/mysql--resource_version_update-increment resource version.sql
… pkg/storage/unified/sql/testdata/postgres--resource_history_insert-insert into resource_history.sql
… pkg/storage/unified/sql/testdata/postgres--resource_read-without_resource_version.sql
… pkg/storage/unified/sql/testdata/postgres--resource_update-single path.sql
… pkg/storage/unified/sql/testdata/postgres--resource_version_get-single path.sql
… pkg/storage/unified/sql/testdata/postgres--resource_version_inc-increment resource version.sql
… pkg/storage/unified/sql/testdata/postgres--resource_version_insert-single path.sql
… pkg/storage/unified/sql/testdata/postgres--resource_version_update-increment resource version.sql