**What is this feature?**

Introduce tenant shards in the resource_version to increase write-throughput.


Before: We serialise the writes per resource.
After: We serialise the writes per resource per shard (has of the namespace). At the expense of the Initial LIST call for cross-namespace queries being done at a recent RV (lagging behind concurrent ongoing transactions).

## Benchmark
Benchmark simulating 10ms lock time latency and 12 concurrent tenants, 1 resource type:
- 😢  Without sharding, we can achieve 90 writes per seconds**
- 😄 With sharding , we get to ~900writes per second

```
# NumberOfShards=1
% GRAFANA_TEST_DB=mysql go  test -benchmem -run=^$ -bench ^BenchmarkWriteEvent$ github.com/grafana/grafana/pkg/storage/unified/sql/test  -count 5 -benchtime=500x
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/storage/unified/sql/test
cpu: Apple M3 Pro
BenchmarkWriteEvent-12               500          14226088 ns/op           58906 B/op       1312 allocs/op
BenchmarkWriteEvent-12               500          14255228 ns/op           58214 B/op       1311 allocs/op
BenchmarkWriteEvent-12               500          14228667 ns/op           58273 B/op       1311 allocs/op
BenchmarkWriteEvent-12               500          14404080 ns/op           58365 B/op       1311 allocs/op
BenchmarkWriteEvent-12               500          14322410 ns/op           58390 B/op       1311 allocs/op
PASS
ok      github.com/grafana/grafana/pkg/storage/unified/sql/test 40.404s

# NumberOfShards=100
 % GRAFANA_TEST_DB=mysql go  test -benchmem -run=^$ -bench ^BenchmarkWriteEvent$ github.com/grafana/grafana/pkg/storage/unified/sql/test  -count 5 -benchtime=500x
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/storage/unified/sql/test
cpu: Apple M3 Pro
BenchmarkWriteEvent-12               500           1190093 ns/op           59133 B/op       1317 allocs/op
BenchmarkWriteEvent-12               500           1178384 ns/op           59536 B/op       1317 allocs/op
BenchmarkWriteEvent-12               500           1991839 ns/op           58545 B/op       1316 allocs/op
BenchmarkWriteEvent-12               500           1169455 ns/op           59318 B/op       1316 allocs/op
BenchmarkWriteEvent-12               500           2987246 ns/op           59001 B/op       1315 allocs/op
PASS
ok      github.com/grafana/grafana/pkg/storage/unified/sql/test 8.049s
```

**Why do we need this feature?**
Elevate the write throughput bottleneck in unistore
 
**Special notes for your reviewer:**

Please check that:
- [ ] It works as expected from a user's perspective.
- [ ] If this is a pre-GA feature, it is behind a feature toggle.
- [ ] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/contribute/release-notes/#how-to-determine-if-content-belongs-in-whats-new), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/contribute/release-notes/) doc.